more progress on all files made more variables and equations for each
