+++ src/oop/items.java
@@ -1,0 +1,17 @@
+package oop;
+import java.util.ArrayList;
+
+import oop.*;
+
+public class items{
+  private String names;
+  private double total;
+  private ArrayList<cart> ordercart; 
+public cart(String name){
+  this.names = name;
+  this.total = 0;
+  ordercart= new ArrayList<cart>();
+}
+
+
+}
